Experiment: see how much space multipliers are taking up
As of the last commit, we're at a utilization of 89%... which suggests
that each oscillator adds ~8% to the utilization.

That leaves quite a bit unaccounted for in the utilization.
The multipliers in the iteration calculation are an obvious suspect, so
I'm taking them out to see what the result is.