Asynchronous resets to appease the simulation gods
The testbench does *not* like the fact that I wasn't explicitly
initializing certain registers on reset. Now, for all of said
registers, it was OK to come up as any value, but that leads
to a bunch of b'xxxx values in the simulation.

So, you know what? Sure! We'll initialize them! Flip-flops with
asynchronous resets aren't much bigger than flip-flops without,
right?